chore : useClinicDeleteModal 제거 (#116)
