+++ src/features/index.ts
@@ -11,7 +11,6 @@ export * from './sign-up/UserInfoStep'
 export * from './clinic/hook/useFormattedVisitDate'
 export * from './clinic/hook/useSelectAlarmTime'
 export * from './clinic/hook/useTagToggle'
-export * from './clinic/service/useClinicDeleteModal'
 export * from './clinic/service/useInsertedClinicForm'
 export * from './clinic/service/useMonthlyScheduleList'
 export * from './clinic/service/useSubmitAddClinicForm'
